Fix handling of brackets '[]' in parameter values (e.g. (#297)
in a description field) and when a parameter name contains
'[]' and a POST is used (e.g. during a DeployVirtualMachine).

CloudStack only uses a specialized encoding for the signature
calculation. It only query encodes the values of the query arguments
and always use '%20' for whitespaces (no '+'). But for the GET url
and for the POST body, we always have to use the normal
query escaping as standardized.

Signed-off-by: stffabi <stffabi@pm.me>