Add support for building snap packages
This branch makes a few changes to support running the exo binary from a
snap package (https://snapcraft.io).

The  biggest diffrence as far as this application is concerned is that
snaps source their data from special folders instead of the usual
dotfiles. The configuration is therefore set to honour $SNAP_USER_COMMON
since the data is per-user, and common to all versions of a release (not
per-release).

Note: this is not meant to be "production ready" yet, a few rough
corners need to be polished first, but this at least allows manual
publication into the "edge channel".