sos: fix TLS certificate issue on Windows
This change introduces a workaround to a bug in the Microsoft Windows
support in Go (https://github.com/golang/go/issues/16736). Some
heavily-chained certificates such as Exoscale SOS API's one are not
recognized as being signed by a trusted Certificate Authority.

Until the bug is fixed upstream, Windows users are required to use the
`--certs-file` flag with a path to a local file containing Exoscale SOS
API's certificate chain to the "exo sos" commands.

Fixes #189.