ssh: fix bug with username-less templates
This change fixes a bug in the SSH-related sections of the CLI, where an
error message was reported when dealing with Compute instances based on
a custom template for which a username is not set in its metadata
(setting a username is not mandatory).

Before:

        $ exo ssh --print my-instance
        error: missing username information in Template "58f032e5-20db-425c-8108-37f0b764cd1a"

After:

        $ exo ssh --print my-instance
        ssh -i "/Users/marc/.exoscale/instances/8bc37224-03af-46c5-9e30-ca4cac62a887/id_rsa" 159.100.242.169

Also, a bit of refactoring in `ssh` related commands.