fix(sos): Calculate appropriate partsize when uploading
Due to the library which renders the progress bar, the AWS sdk cannot
compute the correct partsize. This means that the maximum file size
which can be uploaded by the CLI is 50 GB: 5 MB x 10 000 parts, because
the AWS SDK has an fixed max of 10k parts.

The change here explicitly calculates and sets a partsize, based on the
detected size of the object which is being uploaded.

An exo binary with this change applied managed to upload an object of
250 GB.