sos: add support for setting the object ownership + refactoring (#498)
We add support for setting the object ownership.

As a follow up to this feature, we will add support for versioning. We
will need to touch most object storage related commands and in
anticipation of this we refactored most of the object storage related
code. In particular we separated it into a different package. This
should allow us to write better testing later.

---------

Co-authored-by: Philipp Sauter <philipp.sauter@exoscale.ch>