installation: add a script to install the latest version on Debian and Red Hat-based distros. (#505)
In the discussion of #492 it became apparent that some users are lacking
an easy and automatic way to install the latest version of exo cli. We
add a script to automate this for Debian and Red Hat-based
distributions.

The script requires `sh`, `sha256sum`(to check the checksum), `curl` or
`wget` and either `dpkg`, `yum` or `dnf`. It detects automatically which
one of these commands should be used to download and install the
package.

The script can be run in one command with `curl -fsSL <url to script> |
sh` and apart from invoking `sudo` or `doas` it should not require any
more user interaction.

---------

Co-authored-by: Philipp Sauter <philipp.sauter@exoscale.ch>