Ignore regex checks in `validateDatabaseServiceSettings` (#515)
Fixes the issue with validation error for grafana settings
`smtp.from_address`. It was identified as a regex pattern originating
from Aiven that does not match in Go. Because of this we cannot rely on
the clients side validation of patterns and I've chosen to disable the
feature and rely on server side validation.

```
$ go run . dbaas update mydb --grafana-settings '{"smtp_server":{"host":"example.net","port":587,"username":"person@example.net","password":"mypassword","from_address":"dashboard@example.net"}}'                                      
 ✔ Updating Database Service "mydb"... 0s                                                                                    
```