Don't fetch account info when adding new account (#520)
Remove the call to v1 API (`listAccounts` ) when adding new account. We
don't have the same functionality in v2 to do a switch, but feature does
not provide much value so I decided to remove it. Now it works by always
asking user to manually input all account data, which was a failback
before in case of a restricted key:

```
$ go run . config                                                                        
✔ <Configure a new account>                                                                                            
[+] API Endpoint [https://api.exoscale.com/v1]:                                                                        
[+] API Key [none]: EXOxxx...
[+] Secret Key [none]: ...
[+] Account name [none]: api                                                                                           
[+] Name [none]: api                                                                                                   
✔ at-vie-2 
```