release: adapt AUR release script for signed packages (#541)
# Description
We adapted the 'exoscale-cli' and exoscale-cli-bin' packages on AUR to
verify GPG signatures before installation. The sha256sums list now
contains a 'SKIP' element for the corresponding element in the sources
list that references the signature file. We need to adapt the sed line
that substitutes the checksum to respect this.

## Checklist

* [x] Changelog updated (under *Unreleased* block)
* [x] Testing

## Testing
I tested the release script locally on the adapted PKGBUILD files in
'exoscale-cli' and 'exoscale-cli-bin' and was able to build and install
the packages.

---------

Co-authored-by: Philipp Sauter <philipp.sauter@exoscale.ch>