install script: install from SOS apt repo if possible (#551)
# Description
We install the cli as a debian package through apt from the repository
on SOS, if the version of apt is >= 2.2.0. With version 2.2.0 keyrings
under `/etc/apt/keyrings` became the recommended way to install GPG
keys. Distributions who have older versions of apt may still use the
legacy keyring. On these older distributions we keep installing the
package through dpkg.

The advantage of installing from the SOS repo is that apt will detect
new versions when they are released, so users won't have to manually
install them anymore.

---------

Co-authored-by: Philipp Sauter <philipp.sauter@exoscale.ch>