sks: flag for CSI addon (#572)
We add the flag `--exoscale-csi` to the `exo compute sks create`
command. If set, a cluster with the CSI addon will be created. Unlike
the `CCM` and `metrics-server` addons which have to be explicitly
disabled with a `--no-...` flag, the CSI addon has to be explicitly
enabled because for some users may not need it.