storage: Introduce new `purge` call
Adding `exo storage purge` which removes all objects, versions and
delete markers, from a given bucket with an optional prefix. This
simplifies the bucket deletion process (purge bucket, then delete) when
versioning was enabled since most tools do not support it.

I did not reuse `c.ForEachObject` on purpose since it is reads the
entire contents of the bucket before moving forward. Note that no
parallelism was added in this version, but adding some would greatly
improve the performance of purges.